# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API документация для системы управления лидами и интеграциями",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{schema}://{host}",
            "description": "API Server",
            "variables": {
                "schema": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "http"
                },
                "host": {
                    "default": "localhost"
                }
            }
        }
    ],
    "paths": {
...
        "/api/integrations/{service}/credentials": {
            "get": {
                "tags": [
                    "Integrations"
                ],
                "summary": "Получение учетных данных интеграции",
                "operationId": "50c5ccbe3a857b3c0817475e00e29c9e",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "description": "Тип сервиса интеграции",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение данных"
                    }
                }
            },
            "post": {
                "tags": [
                    "Integrations"
                ],
                "summary": "Добавление учетных данных для интеграции",
                "operationId": "049c4aaa38d9e2467aebce84f6c03ee5",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "description": "Тип сервиса интеграции",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное сохранение учетных данных"
                    },
                    "400": {
                        "description": "Ошибка валидации"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Integrations"
                ],
                "summary": "Удаление учетных данных интеграции",
                "operationId": "a13b4da57be96b5c9ddbf05d9552ad57",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "description": "Тип сервиса интеграции",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное удаление"
                    },
                    "500": {
                        "description": "Ошибка при удалении"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Integrations"
                ],
                "summary": "Обновление учетных данных интеграции",
                "operationId": "34f195e290f87c06a03088f640685969",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "description": "Тип сервиса интеграции",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное обновление данных"
                    },
                    "500": {
                        "description": "Ошибка при обновлении"
                    }
                }
            }
        },
...
    },
...
}